backend/pyproject.toml
[project]
name = "genie-in-the-bottle"
version = "0.1.0"
description = "Scripts for posting AI risk updates to X"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "python-dotenv",
    "tweepy",
    "openai",
    "psycopg2-binary",
    "requests",
    "uvicorn",
]

[tool.uv]
dev-dependencies = []

